Safari-609.1.11.1/LayoutTests/imported/w3c/web-platform-tests/html/webappapis/scripting/processing-model-2/runtime-error-data-url-expected.txt

CONSOLE MESSAGE: line 1: ReferenceError: Can't find variable: undefined_variable

PASS window.onerror - runtime error in <script src=data:...> 
PASS window.onerror - runtime error in <script src=data:...> (column) 
